+++ changelog.md
@@ -18,6 +18,7 @@
 * Add - New mobile menu theme option.
 * Fix - Improving Sensei Course Participants widget behavior.
 * Added - `lsx_header_wrap_after` to allow actions above the .wrap.container
+* Added - `lsx_get_template_part()` allowing 3rd part plugin to overwrite the index.php content template.
 
 ### 2.6.1
 * Fix - Removing `lsx_defer_parsing_of_js` and `preload_css` because they have conflicts with cache plugins.